+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] (YYYY-mm-dd)
+
+- feat: better error message in case of failed raw HTTP requests to Solr
+
 ## 2.0.0 (2024-01-04)
 
 - feat: Support Solr Cloud collection configuration during creation (#3)